pyproject.toml
[tool.poetry]
name = "ci-base"
version = "0.0.0"
description = "Base image for CI/CD workloads"
authors = ["Jairo Llopis <jairo.llopis@tecnativa.com>"]
license = "BOOST-1.0"

[tool.poetry.dependencies]
python = "^3.9"
plumbum = "^1.6.9"

[tool.poetry.dev-dependencies]
pytest-xdist = "^2.1.0"
plumbum = "^1.6.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
